requirements.txt
absl-py>=1.0.0
cycler>=0.11.0
dask>=2021.1.0
decorator>=4.4.0
h5py>=3.1.0
matplotlib>=3.3.0
networkx>=2.6
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
protobuf>=3.19.0
pyparsing>=2.4.0
python-dateutil>=2.8.0
PyWavelets>=1.1.1
scikit-image>=0.18.0
scipy>=1.7.0
tensorboard>=2.8.0
tensorflow>=2.8.0
termcolor>=1.1.0
tfmpl>=1.0.2
toolz>=0.11.0
Werkzeug>=2.0.0
yapf>=0.32.0
